utiliser powf: pow avec un float fait un bug quand retourne valeur < 1E-38
git-svn-id: http://einstein/svn/firmware/FIR-00251-Output/trunk@19124 3e98269d-ff51-e449-abca-491a6567d830

+++ src/SDI/AGC/Drivers/AGC.c
@@ -238,7 +238,7 @@ static float AGC_getAlphaFactor(uint32_t timestamp, float responseTime_sec)
     timestamp_prev = timestamp;
 
     // Return alpha factor for response time of 95%
-    return pow(1.0 - 0.95, deltaT_sec / responseTime_sec);
+    return powf(1.0f - 0.95f, deltaT_sec / responseTime_sec);
 }
